LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/add-two-types.tentative-expected.txt


FAIL Adding two types with different non-null percent hints throws TypeError assert_throws_js: function "() => new CSSMathSum(a, b)" did not throw
FAIL Adding two types with the same nonzero values returns same type assert_equals: length expected 1 but got 0
FAIL Adding two types with empty maps with returns empty map assert_equals: length expected (undefined) undefined but got (number) 0
FAIL Adding a type with percent returns type with percent hint assert_equals: length expected 1 but got 0
FAIL Adding a type with percent 2 returns type with percent hint throws TypeError assert_throws_js: function "() => new CSSMathSum(a, b)" did not throw
FAIL Adding a type with a percent hint returns a type with the percent hint assert_equals: length expected 1 but got 0
FAIL Adding two types with the same percent hint returns a type with the percent hint assert_equals: length expected 1 but got 0
